refactor: each rpc method own file
