+++ src/background/messaging/legacy/legacy-external-message-handler.ts
